Move game statistics and leaderboard commands to a dedicated `GameStatisticsModule`
